requirements.txt
aiohttp==3.6.2
amqp==2.5.2
appdirs==1.4.4
apply-defaults==0.1.4
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
billiard==3.6.3.0
celery==4.4.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==6.7
cryptography==2.8
decorator==4.4.1
distlib==0.3.1
Django==3.0.3
django-appconf==1.0.3
django-autoslug==1.9.7
django-ckeditor==5.9.0
django-colorfield==0.3.2
django-cors-headers==3.2.1
django-cropperjs==0.0.5
django-debug-toolbar==2.2
django-filter==2.2.0
django-grappelli==2.14.2
django-js-asset==1.2.2
django-json-rpc==0.7.2
django-modeltrans==0.4.0
django-mptt==0.11.0
django-rest-auth==0.9.5
django-tabbed-admin==1.0.4
django-thumbnail-maker==0.0.6
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
djangorestframework-recursive==0.1.2
djangorestframework-simplejwt==4.4.0
filelock==3.0.12
funcsigs==1.0.2
future==0.18.2
idna==2.9
idna-ssl==1.1.0
importlib-metadata==1.5.0
jsonrpc==1.2
jsonrpcclient==3.3.5
jsonrpcserver==3.5.6
jsonschema==2.6.0
kombu==4.6.7
multidict==4.7.5
mysql-connector==2.2.9
numpy==1.18.2
pandas==1.1.2
pilkit==2.0
Pillow==7.0.0
psycopg2-binary==2.8.4
pycparser==2.19
PyJWT==1.7.1
pyrsistent==0.15.7
python-dateutil==2.8.1
pytz==2019.3
redis==3.5.0
requests==2.23.0
six==1.14.0
sorl-thumbnail==12.6.3
sqlparse==0.3.0
telepot==12.7
tornado==6.0.3
typing-extensions==3.7.4.1
Unidecode==1.1.1
urllib3==1.25.8
vine==1.3.0
virtualenv==20.0.31
xlrd==1.2.0
yarl==1.4.2
zipp==3.0.0
